Refactor Model files and rearrange Model files.
