+++ FindYourOnlys/Model/Adopt/Home/AdoptDetailContentCategory.swift
@@ -42,7 +42,9 @@ enum AdoptDetailContentCategory: String, CaseIterable {
     func cellForIndexPath(_ indexPath: IndexPath, tableView: UITableView, viewModel: PetViewModel) -> UITableViewCell {
 
         guard
-            let cell = tableView.dequeueReusableCell(withIdentifier: AdoptDetailDecriptionTableViewCell.identifier, for: indexPath) as? AdoptDetailDecriptionTableViewCell
+            let cell = tableView.dequeueReusableCell(
+                withIdentifier: AdoptDetailDecriptionTableViewCell.identifier,
+                for: indexPath) as? AdoptDetailDecriptionTableViewCell
                 
         else { return UITableViewCell() }
