+++ FindYourOnlys/Model/ChatSociety/Add Friend/FriendRequest.swift
@@ -7,6 +7,13 @@
 
 import Foundation
 
+struct FriendRequestList {
+    
+    let type: FriendRequestType
+    
+    var users: [User]
+}
+
 struct FriendRequest: Codable {
     
     var requestUserId: String
@@ -15,3 +22,4 @@ struct FriendRequest: Codable {
     
     var createdTime: TimeInterval
 }
+
