--- FindYourOnlys/Model/ChatSociety/Add Friend/FriendRequestList.swift
@@ -1,15 +1,0 @@
-//
-//  FriendRequestList.swift
-//  FindYourOnlys
-//
-//  Created by 鄭昭韋 on 2022/4/18.
-//
-
-import Foundation
-
-struct FriendRequestList {
-    
-    let type: FriendRequestType
-    
-    var users: [User]
-}
