+++ FindYourOnlys/Model/Error/AuthError.swift
