--- FindYourOnlys/Model/FindPetSociety/Home/PetKind.swift
@@ -1,17 +1,0 @@
-//
-//  PetKind.swift
-//  FindYourOnlys
-//
-//  Created by 鄭昭韋 on 2022/4/18.
-//
-
-import Foundation
-
-enum PetKind: String, CaseIterable {
-    
-    case cat = "貓"
-    
-    case dog = "狗"
-    
-    case others = "其他"
-}
