--- FindYourOnlys/Model/FindPetSociety/Home/PostType.swift
@@ -1,15 +1,0 @@
-//
-//  PostType.swift
-//  FindYourOnlys
-//
-//  Created by 鄭昭韋 on 2022/4/18.
-//
-
-import Foundation
-
-enum PostType: String, CaseIterable {
-    
-    case missing = "遺失"
-    
-    case found = "尋獲"
-}
