--- FindYourOnlys/Model/FindPetSociety/Home/Sex.swift
@@ -1,15 +1,0 @@
-//
-//  Sex.swift
-//  FindYourOnlys
-//
-//  Created by 鄭昭韋 on 2022/4/22.
-//
-
-import Foundation
-
-enum Sex: String, CaseIterable {
-    
-    case male = "公"
-    
-    case female = "母"
-}
